[IMG-234] Refactor Tre/TreImpl to avoid unneeded superclass and interface.
