+++ core-api/src/main/java/org/codice/imaging/nitf/core/tre/Tre.java
@@ -11,52 +11,61 @@
  * is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  *
- **/
+ *
+ */
 package org.codice.imaging.nitf.core.tre;
 
 /**
- Tagged registered extension (TRE).
+ * Tagged registered extension (TRE).
  */
-public interface Tre extends TreGroup, TreEntryList {
-    /**
-     Set the metadata prefix format string.
+public interface Tre extends TreGroup {
 
-     @param mdPrefix the metadata prefix.
+    /**
+     * Set the metadata prefix format string.
+     *
+     * @param mdPrefix the metadata prefix.
      */
     void setPrefix(String mdPrefix);
 
     /**
-     Return the metadata prefix format string.
-
-     @return the metadata prefix.
+     * Return the metadata prefix format string.
+     *
+     * @return the metadata prefix.
      */
     String getPrefix();
 
     /**
-     Set the raw data for this TRE.
-     <p>
-     This is only used for TREs that we couldn't parse.
-
-     @param treDataRaw the raw bytes for the TRE.
+     * Set the raw data for this TRE.
+     * <p>
+     * This is only used for TREs that we couldn't parse.
+     *
+     * @param treDataRaw the raw bytes for the TRE.
      */
     void setRawData(byte[] treDataRaw);
 
     /**
-     Get the raw data for this TRE.
-     <p>
-     This is only used for TREs that we couldn't parse.
-
-     @return the raw bytes for the TRE.
+     * Get the raw data for this TRE.
+     * <p>
+     * This is only used for TREs that we couldn't parse.
+     *
+     * @return the raw bytes for the TRE.
      */
     byte[] getRawData();
 
     /**
      * The source for this TRE.
      *
-     * In this context, the source is the segment header / subheader that the TRE was read from, or where it should be
-     * written to.
+     * In this context, the source is the segment header / subheader that the
+     * TRE was read from, or where it should be written to.
      *
      * @return source for the TRE.
      */
     TreSource getSource();
+
+    /**
+     * Return the name of the entry list key.
+     *
+     * @return the entry list key name
+     */
+    String getName();
 }
